+++ game/util.py
@@ -1,7 +1,10 @@
+# from game.board import Board
 
 class Util():
+
+    def __init__(self, board=None):
+        self.board = board
     
-  
     def update_coordinates(self, orientation, row, col):
         orientation_updates = {"H": (0, 1), "V": (1, 0)}
         row_update, col_update = orientation_updates.get(orientation, (0, 0))
@@ -24,17 +27,31 @@ class Util():
     #     return isinstance(item, str)
 
 
+
+    def split_groups_in_string(self, input_string):
+        word = []
+        i = 0
+        while i < len(input_string):
+            if i < len(input_string) - 1 and (input_string[i] == 'c' and input_string[i + 1] == 'h' or input_string[i] == 'l' and input_string[i + 1] == 'l' or input_string[i] == 'r' and input_string[i + 1] == 'r'):
+                word.append(input_string[i:i+2])
+                i += 2
+            else:
+                word.append(input_string[i])
+                i += 1
+        return word
+    
+
     def get_word_horizontal(self, row, col):
         word = []
-    
+        grid = self.board.grid
         for c in range(col, -1, -1):
-            if self.grid[row][c].has_tile():
-                word.insert(0, self.grid[row][c])
+            if grid[row][c].has_tile():
+                word.insert(0, grid[row][c])
             else:
                 break
-        for c in range(col + 1, len(self.grid[row])):
-            if self.grid[row][c].has_tile():
-                word.append(self.grid[row][c])
+        for c in range(col + 1, len(grid[row])):
+            if grid[row][c].has_tile():
+                word.append(grid[row][c])
             else:
                 break
 
@@ -42,18 +59,19 @@ class Util():
             return False
         return word
     
+
 
     def get_word_vertical(self, row, col):
         word = []
-    
+        grid = self.board.grid
         for r in range(row, -1, -1):
-            if self.grid[r][col].has_tile():
-                word.insert(0, self.grid[r][col])
+            if grid[r][col].has_tile():
+                word.insert(0, grid[r][col])
             else:
                 break
-        for r in range(row + 1, len(self.grid)):
-            if self.grid[r][col].has_tile():
-                word.append(self.grid[r][col])
+        for r in range(row + 1, len(grid)):
+            if grid[r][col].has_tile():
+                word.append(grid[r][col])
             else:
                 break
 
@@ -61,24 +79,25 @@ class Util():
             return False
         return word
     
+
+
     def get_word_left(self, row, col):
         word = []
-        while col >= 0 and self.grid[row][col-1].has_tile():
-            word.insert(0, self.grid[row][col-1])
+        grid = self.board.grid
+        while col >= 0 and grid[row][col-1].has_tile():
+            word.insert(0, grid[row][col-1])
             col -= 1
         return word
     
-
+    
     def get_word_right(self, row, col):
         word = []
-        while col < len(self.grid[row]) and self.grid[row][col + 1].has_tile():
-            word.append(self.grid[row][col + 1])
+        grid = self.board.grid
+        while col < len(grid[row]) and grid[row][col + 1].has_tile():
+            word.append(grid[row][col + 1])
             col += 1
         return word
-    
 
-
-